requirements.txt
BeautifulSoup==3.2.1
Django==1.3.1
Fabric==1.4.0
South==0.7.3
boto==2.2.2
django-analytical==0.12.1
django-annoying==0.7.6
django-compressor==1.1.2
django-extensions==0.7.1
django_facebook==4.0.9
django-ses
django-selector==0.3
django-staticfiles==1.2.1
django-storages==1.1.4
gunicorn==0.13.4
lettuce==0.1.34
lxml==2.3.3
mechanize==0.2.5
nose==1.1.2
pil==1.1.7
psycopg2==2.4.4
pycrypto==2.5
pytz==2011n
rdflib==3.1.0
salad==0.4.3
selenium==2.17.0
six==1.1.0
splinter==0.4.4
ssh==1.7.13
versiontools==1.8.3
wsgiref==0.1.2
zope.event==3.5.1
zope.interface==3.8.0
zope.schema==4.0.1
zope.testbrowser==4.0.2
